// src/app/models/recettes.ts
export class Recettes {
    id: number;
    ordre: number;
    datec: Date;
    prix: number;
    ref: string;
    tel: string;
    etat: number;
    payer: boolean;
    effecter: boolean;
    categorie: number;
    categorieName: string;
    categorieColor: string;
    abonne: number;
    abonneName: string;
    userid: number;
}